pyproject.toml
[tool.poetry]
name = "apibase"
version = "0.1.0"
description = ""
authors = ["user.name <gmail@hdknr.com>"]

[tool.poetry.dependencies]
python = "^3.8"
graphene-django = "^2.10.1"
django-cors-headers = "^3.3.0"
django-filter = "^2.2.0"
djangorestframework = "^3.11.0"
djangorestframework-filters = {git = "https://github.com/philipn/django-rest-framework-filters.git"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
